Added upload and download methods
