+++ autotest-web/src/main/resources/application.properties
@@ -1,2 +1,3 @@
 base.url=http://178.154.246.238:58082/
 auth.yaml=src/main/resources/auth.yaml
+nopict=src/test/resources/files/test.txt
