+++ autotest-web/src/test/resources/features/tk6/6.3_EmployeePageNegativeHrActionEmptyField.feature
@@ -15,11 +15,11 @@
     Если кнопка '<button_emp>' отображается
     Тогда нажать на '<button_emp>'
     Тогда инициализация страницы 'DjangoEmployee'
-    Если кнопка '<button2>' отображается
-    Тогда нажать на '<button2>'
+    Если кнопка '<button>' отображается
+    Тогда нажать на '<button>'
     И поле '<field>' инфоблок с текстом '<text>' присутствует
 
     Примеры:
-      |login |button_emp |button2  |field                      |text                                                                                     |
+      |login |button_emp |button   |field                      |text                                                                                     |
       |hr    |Сотрудники |Выполнить|Предупреждение в заголовке |Чтобы произвести действия над объектами, необходимо их выбрать. Объекты не были изменены.|
