+++ autotest-web/src/test/resources/files/test.txt
@@ -1,0 +1,1 @@
+file for negative tests
